+++ src-tauri/src/utils/app/mod.rs
@@ -1,3 +1,4 @@
 pub mod builder;
 pub mod notification;
 pub mod ui;
+pub mod updater;